+++ .eslintrc
@@ -61,9 +61,6 @@
 		"@typescript-eslint/prefer-for-of": "error",
 		"@typescript-eslint/prefer-includes": "error",
 		"@typescript-eslint/prefer-nullish-coalescing": "error",
-		"@typescript-eslint/prefer-string-starts-ends-with": "error",
-		"@typescript-eslint/consistent-type-imports": ["error", {
-			"prefer": "type-imports"
-		}]
+		"@typescript-eslint/prefer-string-starts-ends-with": "error"
 	}
 }
