+++ src/plugins/utils/tree-clusters.ts
@@ -1,4 +1,4 @@
-import type {
+import {
     ClusterizedFlatTree,
     MetaClusterizedFlatTree,
     ClusterizedFlatTreeNode,
